+++ changelog.md
@@ -1,5 +1,9 @@
 <h3>Changelog</h3>
 
+## 68 - June 19, 2015
+
+* Update HTML5 fix for the old player. `Objects.keys(obj.constructor.prototype)` was skipping the property, so used `for (var k in obj){}`.
+
 ## 67 - June 17, 2015
 
 * Make script slightly compatible with Youtube+